Fix for empty name remote output from irsend list
For some reason when i run irsend list "" "" the first line of remotes is an empty line. I have looked for a faulty conf file in /etc/lirc/* but i have not found any. Without the check if a remote is a string the -1 lookup will give an index out of range error.